feat(storage-decoupling-global-stanza): add global path map and extract removeFencedContent
Create global-path-map.ts mapping all five harness IDs to their
user-global instruction file paths with a resolveGlobalInstructionPath
convenience function that accepts homeDir for test isolation.

Extract removeFencedContent from uninstall/index.ts into
comment-fence.ts as a shared export so both uninstall and the upcoming
global stanza writer can use it without duplicating the logic.
